[frontend] Move GraphTable creation to GraphTableProvider
