+++ frontendGraphTableProvider/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/table/GraphTableProvider.java
@@ -1,33 +1,23 @@
 package com.virtuslab.gitmachete.frontend.ui.api.table;
 
-import com.intellij.openapi.actionSystem.DataProvider;
 import com.intellij.openapi.components.Service;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.wm.ToolWindowId;
-import com.intellij.openapi.wm.ToolWindowManager;
-import io.vavr.control.Option;
-import org.checkerframework.checker.guieffect.qual.UIEffect;
+import lombok.Getter;
 
-import com.virtuslab.gitmachete.frontend.datakeys.DataKeys;
+import com.virtuslab.binding.RuntimeBinding;
 
 @Service
+@Getter
 public final class GraphTableProvider {
-  private final Project project;
+
+  private final BaseGraphTable graphTable;
+
+  private final VcsRootComboBox vcsRootComboBox;
 
   public GraphTableProvider(Project project) {
-    this.project = project;
-  }
-
-  @UIEffect
-  public Option<BaseGraphTable> getGraphTable() {
-    var toolWindow = ToolWindowManager.getInstance(project).getToolWindow(ToolWindowId.VCS);
-    if (toolWindow == null) {
-      return Option.none();
-    }
-
-    var contentManager = toolWindow.getContentManager();
-    var git_machete = contentManager.findContent("Git Machete");
-    var component = (DataProvider) git_machete.getComponent();
-    return Option.of((BaseGraphTable) component.getData(DataKeys.KEY_GRAPH_TABLE.getName()));
+    vcsRootComboBox = new VcsRootComboBox(project);
+    this.graphTable = RuntimeBinding
+        .instantiateSoleImplementingClass(IGraphTableFactory.class)
+        .create(project, vcsRootComboBox);
   }
 }
